+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -231,7 +231,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K9RWC11ZN270TJGGFYPP1JJ4"),
+  _id: z.string().default("01K9SAG3NYM3M6119WE2BY6F89"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -274,7 +274,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01K9RWC11ZN270TJGGFYPP1JJ4"),
+  id: z.string().default("01K9SAG3NYM3M6119WE2BY6F89"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,