+++ packages/orq-rc/RELEASES.md
@@ -4909,3 +4909,13 @@ Based on:
 - [typescript v4.0.0-rc.22] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.22 - packages/orq-rc
+
+## 2025-11-11 11:25:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.653.0 (2.748.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.23] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.23 - packages/orq-rc